Create de Brand criado
