+++ package.json
@@ -33,7 +33,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "nodemon": "^2.0.22",
     "prettier": "^2.8.8",
-    "prisma": "^4.15.0",
+    "prisma": "^4.16.2",
     "sucrase": "^3.32.0"
   }
 }
