+++ src/modules/brand/Controller/findAll.js
@@ -1,0 +1,1 @@
+/* export const read = async (req, res) => {}; */
