+++ src/modules/brand/Controller/index.js
@@ -1,0 +1,5 @@
+export * from './findById.js';
+export * from './findAll.js';
+export * from './excludeById.js';
+export * from './create.js';
+export * from './update.js';
