+++ src/modules/brand/Controller/update.js
@@ -1,0 +1,1 @@
+/* export const update = async (req, res) => {}; */
