+++ src/modules/brand/Dto/example-dto.md
@@ -1,0 +1,1 @@
+A camada DTO serve para serialização de algum objeto, exemplo seria uma serialização de um objeto JSON para o client e sua deserialização quando chegar no client.
