+++ src/modules/brand/Service/createBrandService.js
@@ -1,0 +1,5 @@
+import { createBrandRepo } from "../Repository/createBrandRepo.js";
+
+export const createBrandService = async ({ name }) => { 
+  await createBrandRepo(name)
+};
