--- src/modules/brand/brand.service.js
@@ -1,12 +1,0 @@
-export const validateBrandBeforeCreate = ({ name }) => {
-  try {
-    const brand = db.brand.create({
-      data: {
-        name: name,
-      },
-    });
-    return brand;
-  } catch (error) {
-    throwError('Error creating table car');
-  }
-};
